+++ src/com/art/tech/db/ProductInfoColumn.java
@@ -104,7 +104,7 @@ public class ProductInfoColumn extends DatabaseColumn {
 	    dbHelper.insert(ProductInfoColumn.TABLE_NAME, values);
 	}
 
-	public static int delete(Context c,  int id) {
+	public static int delete(Context c,  long id) {
 		DBHelper dbHelper = DBHelper.getInstance(c);
 		return dbHelper.delete(ImageCacheColumn.TABLE_NAME, id);
 	}
